requirements.txt
streamlit
<<<<<<< HEAD
numpy
tensorflow==2.20.0
librosa
plotly
pandas
matplotlib
scipy
soundfile
audioread
scikit-learn
pillow
=======
tensorflow==2.10.0
numpy
pandas
opencv-python
matplotlib
librosa
soundfile
>>>>>>> 9a4dce49a442d3cd4a2030fc2c8538710ea8d305
